pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama-anvil"
version = "0.1.0"
description = "Batteries-included CLI for Ollama — hardware-aware setup, context compression, HF model import, MCP web search, and agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["ollama", "local-ai", "llm", "agents", "mcp", "huggingface", "gguf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "duckduckgo-search>=6.0.0",
]

[project.optional-dependencies]
tui = ["textual>=0.40.0"]
# Deps for 'anvil models convert' (HF safetensors -> GGUF).
# llama.cpp itself (the convert_hf_to_gguf.py script and llama-quantize
# binary) is installed separately via 'anvil models convert-setup'.
convert = [
    "huggingface_hub>=0.20",
    "torch>=2.0",
    "transformers>=4.40",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20",
]
all = ["ollama-anvil[tui,convert]"]
dev = ["pytest>=7.0.0", "ruff>=0.2.0", "ollama-anvil[all]"]

[project.scripts]
anvil = "anvil.cli:main"

[project.urls]
Homepage = "https://github.com/hariravichandran/ollama-anvil"
Documentation = "https://github.com/hariravichandran/ollama-anvil/tree/main/docs"
Repository = "https://github.com/hariravichandran/ollama-anvil"
Issues = "https://github.com/hariravichandran/ollama-anvil/issues"

[tool.setuptools.packages.find]
include = ["anvil*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",
    "E4",
    "E7",
    "E9",
    "W",
    "I",
    "UP",
    "B",
    "SIM",
]
ignore = [
    "E501",
    "SIM108",
    "SIM114",
]

[tool.ruff.lint.isort]
known-first-party = ["anvil"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "integration: marks tests that require a running Ollama server (deselect with '-m \"not integration\"')",
]
